Do postImage component (frontend), upload image API (backend), writeArticle API (backend)
