+++ backend/package.json
@@ -19,9 +19,12 @@
     "ejs": "^3.1.6",
     "express": "^4.17.1",
     "express-validator": "^6.13.0",
+    "gridfs-stream": "^1.1.1",
     "jsonwebtoken": "^8.5.1",
     "mongoose": "^6.0.12",
     "multer": "^1.4.3",
+    "multer-gridfs-storage": "^5.0.2",
+    "router": "^1.3.6",
     "validator": "^13.7.0"
   },
   "devDependencies": {
