+++ backend/request.http
@@ -1,18 +1,22 @@
 
 ###
-GET http://localhost:9000/getPosts 
+GET http://localhost:9000/getPost
+
+###
+GET http://localhost:9000/posts/getPost
+
 
 ###
 
-PUT http://localhost:9000/article
+PUT http://localhost:9000/posts/writePost
 Content-Type: application/json
 	
 	{	
 		
 		
-		"email": "gentlesoul@gazaphili.com",
-		"title": "Worm birds raising storms",
-		"content": "PRAGUE (Reuters) -The Czech Republic reported 27,717 new coronavirus cases for Thursday, the highest daily tally in the country of 10.7 million since the pandemic started, Health Ministry data showed on Friday. The data showed the epidemic gaining pace with 131,731 cases detected over past seven days, compared to 73,200 cases reported for all of October.",
+		"email":"zhonglei@gmail.com",
+		"title": "The rustle in the mount Part III",
+		"content": "They took my scarf and my staff. They robbed me of my dignity and thought that would emasciate my humanity. But I rose up like the proverbial phoneix, and like a lark sang to the heavens the perreniality of my humanity. The data showed the epidemic gaining pace with 131,731 cases detected over past seven days, compared to 73,200 cases reported for all of October.",
 		"author": "Zhang Licun"
 		
 				
@@ -20,23 +24,23 @@ Content-Type: application/json
 
 
 ###
-POST http://localhost:9000/login
+POST http://localhost:9000/users/login
 Content-Type: application/json
 
 	{  
 	"password": "12345678", 
-	"email": "gentlesoul@gazaphili.com"
+	"email": "zhonglei@gmail.com"
     
 
 	}
 
 ###
-GET http://localhost:9000/users/61a0a72010d9efe46b27591e
+GET http://localhost:9000/users/users/61a0a72010d9efe46b27591e
 
 
 
 ###
-POST http://localhost:9000/register
+POST http://localhost:9000/users/register
 Content-Type: application/json
 
 {
