+++ backend/server.js
@@ -1,12 +1,12 @@
 import express from 'express';
 import cors from 'cors';
 import dotenv from "dotenv";
-import jwt from "jsonwebtoken";
 import { connect } from './libs/database.js';
-import User from "./models/UserProfile.js";
-import { messageRules } from "./validation/messageValidation.js";
+import users from "./routes/userRoutes.js"
+import posts from "./routes/postRoutes.js"
+import uploads from "./routes/uploadsRoutes.js"
 import multer, { diskStorage } from 'multer';
-
+import User from "./models/UserProfile.js";
 //connect NodeJS to database
 dotenv.config();
 await connect()
@@ -19,179 +19,23 @@ app.use(express.static('User'));
 app.use(express.urlencoded({ extended: true }));
 //=============== 
 
-// Register New User ==
+// USER ROUTER HANDLES ALL USER RELATED REQUESTS
+app.use("/users", users);
 
-app.post("/register", async (req, res) => {
-	const user = await User.register(req.body);
+//POSTS ROUTER HANDLES ALL POSTS RELATED REQUESTS
+app.use("/posts", posts);
+//===========
 
-	if (!user) {
-		return res.status(400).json({ success: false });
-	}
-
-	res.status(201).json({ success: true, user: user });
-
-});
-
-// Login User with password and email
-app.post("/login", async (req, res) => {
-
-	const user = await User.login(req.body);
-
-	if (!user) {
-		return res.status(400).json({ user });
-
-	}
-
-	// create JWT token
-	const token = jwt.sign({ _id: user._id }, process.env.SECRET);
-
-	res.json({ user, token })
-	console.log(user);
-})
-
-// JWT validation middleware
-const checkLogin = (req, res, next) => {
-	const rawJWTHeader = req.headers.authorization;
-
-	if (!rawJWTHeader) {
-		return res.sendStatus(401);
-	}
-	const token = rawJWTHeader.slice(7)
-	jwt.verify(token, process.env.SECRET, function (err, decoded) {
-		if (err) {
-			console.log("Cannot verify JWT", err.message);
-			return res.sendStatus(401);
-
-		}
-		next()
-	})
-};
-
-//REQUEST HAndling Middlewares
-
-const validate = (rules) => {
-	const middlewares = rules;
-	middlewares.push((req, res, next) => {
-		const errors = validationResult(req);
-		if (errors.isEmpty()) {
-			return (next())
-
-		}
-		res.status(400).json({
-			errors: errors.array()
-		})
-
-	})
-	return middlewares;
-
-}
-
-// Blog Post endpoint
-
-/* Create POSTS by one user */
-app.put('/article', async (req, res, next) => {
-	const data = req.body
-	console.log("FRONTEND DATA: ", req.body)
-	console.log("POST TITLE", data.title);
-	const storePost = await User.findOne({ email: req.body.email })
-
-	if (storePost) {
-		storePost.posts.push({ title: req.body.title, content: req.body.content, author: req.body.author })
-		await storePost.save()
-		return res.status(201).json({ content: storePost.posts })
-
-	}
-	console.log("I store your posts: ", storePost);
-	res.json({ error: "User cannot be updated" })
-});
+ app.use("/uploads" , uploads)
+app.use(express.static('images'))
 
 
-// GET MESSAGE ARCHIVE 
-
-//===== Get messages of all users
-app.get('/getPosts', async (req, res, next) => {
-	const userPosts = await User.find({})
-
-	for (let index = 0; index < userPosts.length; index++) {
-		const element = userPosts[index];
-		console.log("ELEMENT HERE: ", element.posts[0].author, element.posts)
-		return res.send(element.posts)
-	}
-});
-// END GET POST
-app.get("/users/:_id", async (req, res) => {
-
-	try {
-		let user = await User.findById(req.params._id);
-		console.log('USER_PROFILE: ', user)
-		const data = user
-		res.status(200)
-		return res.send(data);
-	} catch (err) {
-		res.status(500).json({
-			Message: err.message
-
-		})
-	}
-})
-
-//========
-
-// GET USER PROFILE
-
-app.get("/users/:_id", async (req, res) => {
-
-	try {
-		let user = await User.findById(req.params._id);
-		console.log('USER_PROFILE: ', user)
-		const data = user
-		res.status(200)
-		return res.send(data);
-	} catch (err) {
-
-		res.status(500).json({
-			Message: err.message
-
-		})
-
-	}
-})
-
-//==========================
 // Final error check middleware
 app.use((req, res) => {
 	res.status(404);
 	res.json({ error: "Resource not found 😥" });
 });
 //===================
-
-
-// STORE IMAGES LOCALLY: 
-const storageEngine = diskStorage({
-	destination: (req, file, cb) => {
-		cb(null, './images')
-	},
-	filename: (req, file, cb) => {
-		cb(null, Date.now() + '_' + file.originalname)
-	}
-});
-const upload = multer({ storage: storageEngine })
-
-
-
-// IMAGE POST ENDPOINT
-
-// For Single image upload
-app.post('/single', upload.single("myImage"), async (req, res, next) => {
-	const data = req.file
-	console.log(data)
-
-
-
-})
-app.use(express.static('ProfileImg'))
-
-// =================
